requirements.txt
Django>=4.2,<5.0
python-decouple>=3.8
whitenoise>=6.6
django-ratelimit>=4.1
